Only test OOP builds
